+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.1
 - Added new preview in appstore
-- Added new user interface and administration section.
+- Added new user interface and administration section
 - Updated dependencies
+- Removed support of ignored file configuration from web interface
 # 1.0.0
-Added the first version of the app which is able to find duplicates. Working on NC26, NC27, NC28.
+Added the first version of the app which is able to find duplicates. Working on NC26, NC27, NC28